\name{wbm}
\alias{wbm}
\docType{data}
\title{
Whole Bone Marrow
}
\description{
Whole bone marrow, pre-treatment samples
}
\usage{data("wbm")}
\format{
  A data frame with 17480 observations on the following 440 variables.
  \describe{
    \item{\code{BI12.000203.CEL}}{a numeric vector}
    \item{\code{BI12.000210.CEL}}{a numeric vector}
    \item{\code{BI12.000512.CEL}}{a numeric vector}
    \item{\code{BI12.000513.CEL}}{a numeric vector}
    \item{\code{BI13.000437.CEL}}{a numeric vector}
    \item{\code{BI14.000134.CEL}}{a numeric vector}
    \item{\code{BI14.000195.CEL}}{a numeric vector}
    \item{\code{BI14.000290.CEL}}{a numeric vector}
    \item{\code{BI14.000292.CEL}}{a numeric vector}
    \item{\code{BI14.000293.CEL}}{a numeric vector}
    \item{\code{BI14.000301.CEL}}{a numeric vector}
    \item{\code{BI14.000318.CEL}}{a numeric vector}
    \item{\code{BI14.000353.CEL}}{a numeric vector}
    \item{\code{BI14.000355.CEL}}{a numeric vector}
    \item{\code{BI14.000362.CEL}}{a numeric vector}
    \item{\code{BI14.000365.CEL}}{a numeric vector}
    \item{\code{BI14.000367.CEL}}{a numeric vector}
    \item{\code{BI14.000368.CEL}}{a numeric vector}
    \item{\code{BI14.000369.CEL}}{a numeric vector}
    \item{\code{BI14.000370.CEL}}{a numeric vector}
    \item{\code{BI14.000371.CEL}}{a numeric vector}
    \item{\code{BI14.000372.CEL}}{a numeric vector}
    \item{\code{BI14.000373.CEL}}{a numeric vector}
    \item{\code{BI14.000374.CEL}}{a numeric vector}
    \item{\code{BI14.000375.CEL}}{a numeric vector}
    \item{\code{BI14.000376.CEL}}{a numeric vector}
    \item{\code{BI14.000377.CEL}}{a numeric vector}
    \item{\code{BI14.000379.CEL}}{a numeric vector}
    \item{\code{BI14.000380.CEL}}{a numeric vector}
    \item{\code{BI14.000381.CEL}}{a numeric vector}
    \item{\code{BI14.000382.CEL}}{a numeric vector}
    \item{\code{BI14.000383.CEL}}{a numeric vector}
    \item{\code{BI14.000384.CEL}}{a numeric vector}
    \item{\code{BI14.000385.CEL}}{a numeric vector}
    \item{\code{BI14.000386.CEL}}{a numeric vector}
    \item{\code{BI14.000387.CEL}}{a numeric vector}
    \item{\code{BI14.000392.CEL}}{a numeric vector}
    \item{\code{BI14.000393.CEL}}{a numeric vector}
    \item{\code{BI14.000394.CEL}}{a numeric vector}
    \item{\code{BI14.000395.CEL}}{a numeric vector}
    \item{\code{BI14.000396.CEL}}{a numeric vector}
    \item{\code{BI14.000397.CEL}}{a numeric vector}
    \item{\code{BI14.000398.CEL}}{a numeric vector}
    \item{\code{BI14.000399.CEL}}{a numeric vector}
    \item{\code{BI14.000400.CEL}}{a numeric vector}
    \item{\code{BI14.000404.CEL}}{a numeric vector}
    \item{\code{BI14.000406.CEL}}{a numeric vector}
    \item{\code{BI14.000418.CEL}}{a numeric vector}
    \item{\code{BI14.000423.CEL}}{a numeric vector}
    \item{\code{BI14.000424.CEL}}{a numeric vector}
    \item{\code{BI14.000425.CEL}}{a numeric vector}
    \item{\code{BI14.000428.CEL}}{a numeric vector}
    \item{\code{BI14.000431.CEL}}{a numeric vector}
    \item{\code{BI14.000432.CEL}}{a numeric vector}
    \item{\code{BI14.000437.CEL}}{a numeric vector}
    \item{\code{BI14.000468.CEL}}{a numeric vector}
    \item{\code{BI14.000481.CEL}}{a numeric vector}
    \item{\code{BI14.000488.CEL}}{a numeric vector}
    \item{\code{BI14.000489.CEL}}{a numeric vector}
    \item{\code{BI14.000490.CEL}}{a numeric vector}
    \item{\code{BI14.000492.CEL}}{a numeric vector}
    \item{\code{BI14.000510.CEL}}{a numeric vector}
    \item{\code{BI14.000568.CEL}}{a numeric vector}
    \item{\code{BI14.000585.CEL}}{a numeric vector}
    \item{\code{BI14.000587.CEL}}{a numeric vector}
    \item{\code{BI14.000604.CEL}}{a numeric vector}
    \item{\code{BI14.000608.CEL}}{a numeric vector}
    \item{\code{BI14.000623.CEL}}{a numeric vector}
    \item{\code{BI14.000677.CEL}}{a numeric vector}
    \item{\code{BI14.000712.CEL}}{a numeric vector}
    \item{\code{BI14.000717.CEL}}{a numeric vector}
    \item{\code{BI14.000725.CEL}}{a numeric vector}
    \item{\code{P0462.03.BS6067.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0521.00.BS1791.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0990.00.BS1465.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0993.01.BS1527.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P0995.03.BS1694.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1003.05.BS1581.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1010.00.BS1592.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1016.03.BS4747.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1023.02.BS1533.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1027.07.BS1611.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1053.07.BS1189.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1063.05.BS1489.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1064.00.BS1502.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1068.00.BS1688.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1069.00.BS1807.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1070.03.BS1946.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1077.04.BS1984.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1079.04.BS2012.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1081.07.BS2013.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1086.05.BS1873.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1088.04.BS1985.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1093.04.BS2217.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1100.00.BS2287.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1101.05.BS2338.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1112.03.BS2090.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1113.03.BS2092.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1117.03.BS2133.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1119.00.BS2385.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1120.03.BS1293.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1130.01.BS2302.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1132.01.BS2383.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1138.03.BS2236.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1141.04.BS2111.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1142.03.BS2131.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1143.04.BS1890.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1144.04.BS2162.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1145.00.BS2271.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1158.01.BS1627.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1164.00.BS2419.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1165.02.BS2327.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1166.00.BS2357.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1169.04.BS2505.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1171.00.BS2531.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1173.03.BS2435.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1174.00.BS2558.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1175.00.BS2160.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1176.00.BS2578.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1177.04.BS2660.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1178.03.BS2524.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1182.00.BS2610.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1183.02.BS2259.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1184.02.BS2658.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1186.00.BS2601.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1187.02.BS2688.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1188.00.BS2690.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1189.05.BS2699.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1195.02.BS2632.U133Plus.2_2.cel}}{a numeric vector}
    \item{\code{P1196.00.BS2288.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1224.03.BS2977.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1225.00.BS2981.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1228.02.BS3007.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1231.03.BS2608.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1232.02.BS2672.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1239.05.BS2918.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1248.04.BS2916.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1259.02.BS3102.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1263.04.BS2878.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1271.02.BS3149.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1291.01.BS3367.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1296.07.BS3323.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1305.04.BS3474.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1307.03.BS3498.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1314.04.BS3295.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1321.04.BS3368.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1323.00.BS3577.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1331.01.BS3722.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1336.01.BS3696.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1337.02.BS3790.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1338.02.BS3856.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1339.02.BS3884.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1341.02.BS3678.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1350.02.BS4026.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1354.04.BS4009.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1359.00.BS4036.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1363.04.BS3933.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1366.04.BS3840.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1367.02.BS3942.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1371.02.BS4201.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1372.02.BS4047.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1375.02.BS3600.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1379.00.BS3822.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1380.00.BS4012.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1383.04.BS4205.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1384.03.BS4352.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1389.02.BS3438.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1390.02.BS4066.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1391.02.BS4264.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1402.02.BS4590.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1405.02.BS4588.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1407.02.BS4696.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1408.02.BS4661.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1409.02.BS4560.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1412.00.BS4825.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1424.01.BS4756.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1425.01.BS4760.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1432.02.BS5482.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1434.03.BS4955.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1437.04.BS4828.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1441.04.BS5083.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1442.00.BS5108.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1445.02.BS5096.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1446.02.BS4814.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1448.04.BS5055.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1451.02.BS5042.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1452.00.BS4813.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1453.03.BS4848.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1458.02.BS4771.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1459.01.BS4906.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1463.02.BS5288.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1465.03.BS5756.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1483.01.BS5775.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1492.02.BS5605.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1502.02.BS5939.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1503.02.BS4359.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1529.03.BS6081.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1550.03.BS5597.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1571.04.BS5516.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1579.02.BS6355.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1580.04.BS5745.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1581.04.BS6023.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1583.02.BS6181.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1587.03.BS6439.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1591.04.BS5742.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1594.02.BS5871.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1595.04.BS5940.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1596.02.BS5862.U133Plus.2_2.cel}}{a numeric vector}
    \item{\code{P1599.03.BS6194.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1603.02.BS6450.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1605.02.BS6443.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1606.02.BS6521.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1628.03.BS6893.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1632.11.BC12719.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P1634.03.BS8179.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1644.02.BS6827.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1646.02.BS6852.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1670.02.BS6762.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1691.03.BS7560.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1693.03.BS7566.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1694.04.BS7515.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1707.00.BS7568.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1735.03.BS7831.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1748.02.BS7643.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1759.02.BS6906.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P1773.01.BS6008.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1774.01.BS7846.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1775.01.BS7978.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1778.01.BS7341.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1779.02.BS7783.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1780.01.BS5094.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1781.01.BS7871.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1782.01.BS7432.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1783.01.BS7873.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1784.01.BS7829.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1788.01.BS7474.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1789.01.BS7516.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1790.01.BS7745.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1803.01.BS7748.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1805.01.BS7494.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1905.04.BS8914.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1908.03.BS8949.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1919.03.BS9064.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1939.01.BS8712.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1967.00.BS9226.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1969.07.BC0940.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1975.05.BS9095.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1978.03.BS9505.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1979.04.BS9577.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1983.00.BS9310.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1984.00.BS9133.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1986.03.BC0428.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P1989.03.BS9065.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2001.07.BS8946.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2007.02.BC0293.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2011.01.BS9507.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2013.03.BS9451.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2015.03.BS9845.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2018.03.BS9735.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2022.02.BS9891.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2032.02.BS9775.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2039.04.BS9929.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2044.03.BS9549.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2045.03.BS9132.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2061.03.BS9774.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2066.03.BC0214.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2068.04.BC0461.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2079.03.BC1539.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2086.00.BC0443.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2088.04.BC0557.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2097.00.BC0356.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2100.06.BC0279.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2102.05.BC0673.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2106.02.BC0151.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2107.03.BC0316.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2120.02.BC0153.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2121.02.BC0077.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2122.03.BC0911.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2129.02.BC0667.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2134.01.BC0342.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2135.00.BS9558.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2146.01.BS9995.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2149.03.BC1126.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2152.02.BC1040.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2154.03.BC0802.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2155.00.BC1076.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2166.00.BC1157.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2175.01.BS5539.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2176.02.BC1276.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2177.00.BC1314.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2181.02.BC1404.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2187.00.BC0585.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2191.02.BC1243.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2197.02.BC1278.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2199.02.BC0609.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2204.02.BC1457.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2213.01.BC1553.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2224.05.BC1556.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2229.01.BC1600.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2231.01.BC0554.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2232.01.BC1617.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2238.02.BC1155.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2240.02.BC0518.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2243.04.BC1690.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2252.01.BC1727.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2258.01.BC1890.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2259.01.BC1872.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2263.01.BC1904.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2275.03.BC2068.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2278.06.BC2098.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2283.04.BC2110.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2289.00.BC1939.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2296.00.BC1755.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2304.00.BC1817.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2308.04.BC2148.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2317.07.BC2251.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2346.00.BC2524.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2355.05.BC2595.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2361.02.BC2612.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2366.08.BC2696.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2382.01.BC9602.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2407.01.BC3154.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2413.09.BC2881.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2469.09.BC3718.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2471.08.BC3503.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2480.04.BC3790.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2485.03.BC3797.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2494.08.BC3279.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2513.11.BC4173.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2516.04.BC3989.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2525.07.BC6900B.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2537.04.BC3971.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2548.04.BC4287.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2551.04.BC4180.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2576.03.BC4262.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2584.04.BC4410.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2587.03.BC4411.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2589.03.BC4686.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2594.06.BC7498.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2596.02.BC3988.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2597.02.BC3793.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2598.07.BC5395.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2602.02.BC4615.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2642.04.BC4853.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2652.04.BC4898.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2658.02.BC4308.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2693.05.BC5344.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2696.03.BC3742.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2707.10.BC5959.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2708.05.BC5956.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2726.04.BC6063.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2751.01.BC4100.U133Plus.2.cel}}{a numeric vector}
    \item{\code{P2755.05.BC6315.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2763.07.BC6406.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2778.03.BC6500.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2788.04.BC6577.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2796.02.BC6478.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2799.04.BC6687.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2804.04.BC6749.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2814.04.BC6766.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2824.04.BC6694B.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2838.06.BC7029.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2840.07.BC7003.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2841.05.BC7027.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2867.06.BC7190.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2898.04.BC7504B.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P2987.04.BC8286.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3823.02.BC15970.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3931.01.BC17214.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P3945.01.BC17297.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4003.03.BC18194.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4383.01.BC21028.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4384.01.BC20105.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4385.01.BC21630.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4386.01.BC21557.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4387.01.BC20418.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4389.01.BC22465.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4390.01.BC20409.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4391.01.BC22135.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4401.01.BC20355.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4402.01.BC20352.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4403.01.BC21056.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4404.01.BC22905.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4405.01.BC22771.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4406.01.BC21719.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4407.01.BC22089.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4408.01.BC22599.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4412.01.BC13372.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4413.01.BC10983.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4414.01.BC15085.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4415.01.bc14973.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4416.01.BC12508.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4417.01.BC10913.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4418.01.BC10761.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4420.01.BC10779.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4421.01.BC13546.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4422.01.BC9858.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4423.01.BC13447.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4427.01.BC9958.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4428.01.BC13521.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4429.01.BC15370.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4430.01.BC10702.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4431.01.BC13350.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4432.01.BC11488.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4433.01.BC11426.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4434.01.BC14365.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4435.01.BC12612.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4437.01.BC12293.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4438.01.BC10930.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4439.01.BC18826.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4440.01.BC18812.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4442.01.BC18713.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4445.01.BC18054.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4446.01.BC18052.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4449.01.BC17584.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4451.01.bC19676.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4452.01.BC19650.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4455.01.BC19503.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4456.01.BC19451.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4457.01.BC19436.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4458.01.BC19300.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4460.01.BC19044.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{P4493.01.BC15097.U133Plus.2.CEL}}{a numeric vector}
    \item{\code{RE12.000060.CEL}}{a numeric vector}
    \item{\code{RE12.000061.CEL}}{a numeric vector}
    \item{\code{RE12.000137.CEL}}{a numeric vector}
    \item{\code{RE12.000138.CEL}}{a numeric vector}
    \item{\code{RE12.000141.CEL}}{a numeric vector}
    \item{\code{RE12.000145.CEL}}{a numeric vector}
    \item{\code{RE12.000146.CEL}}{a numeric vector}
    \item{\code{RE12.000149.CEL}}{a numeric vector}
    \item{\code{RE12.000150.CEL}}{a numeric vector}
    \item{\code{RE12.000151.CEL}}{a numeric vector}
    \item{\code{RE12.000155.CEL}}{a numeric vector}
    \item{\code{RE12.000229.CEL}}{a numeric vector}
    \item{\code{RE12.000231.CEL}}{a numeric vector}
    \item{\code{RE12.000236.CEL}}{a numeric vector}
    \item{\code{RE12.000240.CEL}}{a numeric vector}
    \item{\code{RE12.000248.CEL}}{a numeric vector}
    \item{\code{RE12.000254.CEL}}{a numeric vector}
  }
}
\details{
More information available in the publication
}
\source{
Danziger et al. (2019) Identifying a High-risk Cellular Signature in the Multiple Myeloma Bone Marrow Microenvironment
}
\references{
Danziger et al. (2019) Identifying a High-risk Cellular Signature in the Multiple Myeloma Bone Marrow Microenvironment
}
\examples{
data(wbm)
## maybe str(wbm) ; plot(wbm) ...
}
\keyword{datasets}
